Align some `dra` files
