+++ config/splat.pspeu.dra.yaml
@@ -68,6 +68,7 @@ segments:
       - [0x23AD8, c, dra_psp/23AD8]
       - [0x23C18, c, dra/5D5BC] # O0
       - [0x26948, c, dra_psp/26948]
+      - [0x274E8, c, dra_psp/274E8]
       - [0x27FD8, c, dra/692E8]
       - [0x2BF60, c, dra/71830]
       - [0x35238, c, dra/70394]
@@ -144,7 +145,7 @@ segments:
       - [0xEE600, .rodata, dra_psp/1DCE0]
       - [0xEE628, .rodata, dra_psp/1DDC0]
       - [0xEE680, .rodata, dra/5D5BC]
-      - [0xEE6C8, .rodata, dra_psp/26948]
+      - [0xEE6C8, .rodata, dra_psp/274E8]
       - [0xEE708, .rodata, dra/692E8]
       - [0xEE7B0, .rodata, dra/71830]
       - [0xEE840, .rodata, dra/704D0]
