+++ config/splat.us.dra.yaml
@@ -74,7 +74,7 @@ segments:
       - [0x1C18, ci4, memcard_14, 16, 48]
       - [0x1D98, ci4, memcard_15, 16, 48]
       - [0x1F18, .data, save_mgr]
-      - [0x21B8, .data, 4AEA4] # D_800A21B8
+      - [0x21B8, .data, 4B758] # D_800A21B8
       - [0x2438, .data, 4DA70] # D_800A2438
       - [0x243C, .data, demo] # D_800A243C
       - [0x245C, .data, 5087C] # D_800A245C
@@ -249,6 +249,8 @@ segments:
       - [0x48EE4, c, save_mgr]
       - [0x4A538, c, 4A538]
       - [0x4AEA4, c, 4AEA4]
+      - [0x4B758, c, 4B758]
+      - [0x4CE2C, c, 4CE2C]
       - [0x4DA70, c, 4DA70]
       - [0x4F45C, c, collider]
       - [0x4FBF8, c, demo]
